Add files via upload
One file contains the same ABM code but with the rewards changed to encourage full investment. Another file contains the same ABM code but with a lower minimum endowment. Another file contains the same ABM code but ignoring the influence of the number of analysts on the capacity of a VC firm to advise startups or make new investments